build gRPC automatically with cmake

+++ src/libReadUntil/ReadUntilClient.hpp
@@ -6,6 +6,9 @@
  */
 
 //#include <minknow>
+#include <string>
+#include <sstream>
+#include <grpcpp/grpcpp.h>
 
 #ifndef LIBREADUNTIL_READUNTILCLIENT_HPP_
 #define LIBREADUNTIL_READUNTILCLIENT_HPP_
@@ -15,7 +18,7 @@ class ReadUntilClient
 
 	private:
 		std::string mk_host{"127.0.0.1"};
-		uint8_t mk_port{8002};
+		uint16_t mk_port{8002};
 
 	public:
 		ReadUntilClient();
